+++ R/data_inclusionLevels.R
@@ -1,18 +1,43 @@
 #' List alternative splicing annotations
 #'
+#' @param species Character: filter results by species (regular expression)
+#' @param assembly Character: filter results by assembly (regular expression)
+#' @param date Character: filter results by date (regular expression)
+#'
 #' @family functions for PSI quantification
 #' @return Named character vector with splicing annotation names
+#' 
+#' @importFrom data.table data.table
+#' @importFrom R.utils capitalize
 #' @export
 #'
 #' @examples
-#' listSplicingAnnotations()
-listSplicingAnnotations <- function() {
-    c("Human hg19/GRCh37 (2017-10-20)"=
-          "annotationHub_alternativeSplicingEvents.hg19_V2.rda",
-      "Human hg19/GRCh37 (2016-10-11)"=
-          "annotationHub_alternativeSplicingEvents.hg19.rda",
-      "Human hg38 (2018-04-30)"=
-          "annotationHub_alternativeSplicingEvents.hg38_V2.rda")
+#' listSplicingAnnotations() # Return all alternative splicing annotations
+#' listSplicingAnnotations(assembly="hg19") # Search for hg19 annotation
+#' listSplicingAnnotations(assembly="hg38") # Search for hg38 annotation
+#' listSplicingAnnotations(date="201(7|8)") # Search for 2017 or 2018 annotation
+listSplicingAnnotations <- function(species=NULL, assembly=NULL, date=NULL) {
+    df <- data.table(
+        species=c("human", "human", "human"),
+        assembly=c("hg19/GRCh37", "hg19/GRCh37", "hg38"),
+        date=c("2017-10-20", "2016-10-11", "2018-04-30"),
+        filename=c("annotationHub_alternativeSplicingEvents.hg19_V2.rda",
+                   "annotationHub_alternativeSplicingEvents.hg19.rda",
+                   "annotationHub_alternativeSplicingEvents.hg38_V2.rda"))
+    
+    filter <- function(data, col, value) {
+        if (!is.null(value)) {
+            data <- data[grep(value, data[[col]], ignore.case=TRUE)]
+        }
+        return(data)
+    }
+    df <- filter(df, "species", species)
+    df <- filter(df, "assembly", assembly)
+    df <- filter(df, "date", date)
+    
+    ns <- sprintf("%s %s (%s)", capitalize(df$species), df$assembly, df$date)
+    res <- setNames(df$filename, ns)
+    return(res)
 }
 
 #' List alternative splicing annotation files available, as well as custom